+++ presentation/src/main/java/com/anytypeio/anytype/presentation/page/editor/Command.kt
@@ -66,5 +66,7 @@ sealed class Command {
 
     object OpenDocumentMenu : Command()
 
+    object OpenProfileMenu : Command()
+
     object AlertDialog : Command()
 }